Add get_files as a package  (suggest not to merge)
I have added the setup.py and copy get_files into ___init__.py as the following "http://python-packaging.readthedocs.io/en/latest/minimal.html" instructed, but if we want to upload it as a package to PyPI, we have to have the valid authentication information. Or would you like to create individual python package in another way? Thank you!